requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==1.4.0
    # via
    #   tensorboard
    #   tensorflow
altair==4.2.2
    # via streamlit
astunparse==1.6.3
    # via tensorflow
attrs==22.2.0
    # via jsonschema
blinker==1.5
    # via streamlit
cachetools==5.3.0
    # via
    #   google-auth
    #   streamlit
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via streamlit
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
decorator==5.1.1
    # via validators
entrypoints==0.4
    # via altair
flatbuffers==23.3.3
    # via tensorflow
fonttools==4.39.0
    # via matplotlib
gast==0.4.0
    # via tensorflow
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via streamlit
google-auth==2.16.2
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.51.3
    # via
    #   tensorboard
    #   tensorflow
h5py==3.8.0
    # via tensorflow
idna==3.4
    # via requests
importlib-metadata==6.0.0
    # via streamlit
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
jsonschema==4.17.3
    # via altair
keras==2.11.0
    # via
    #   -r requirements.in
    #   tensorflow
kiwisolver==1.4.4
    # via matplotlib
libclang==15.0.6.1
    # via tensorflow
markdown==3.4.1
    # via tensorboard
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.7.1
    # via
    #   -r requirements.in
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
numpy==1.24.2
    # via
    #   -r requirements.in
    #   altair
    #   contourpy
    #   h5py
    #   matplotlib
    #   opt-einsum
    #   pandas
    #   pyarrow
    #   pydeck
    #   seaborn
    #   streamlit
    #   tensorboard
    #   tensorflow
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==23.0
    # via
    #   matplotlib
    #   streamlit
    #   tensorflow
pandas==1.5.3
    # via
    #   -r requirements.in
    #   altair
    #   seaborn
    #   streamlit
pillow==9.4.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   streamlit
protobuf==3.19.6
    # via
    #   streamlit
    #   tensorboard
    #   tensorflow
pyarrow==11.0.0
    # via streamlit
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydeck==0.8.0
    # via streamlit
pygments==2.14.0
    # via rich
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via matplotlib
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   streamlit
pytz==2022.7.1
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.28.2
    # via
    #   requests-oauthlib
    #   streamlit
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rich==13.3.2
    # via streamlit
rsa==4.9
    # via google-auth
seaborn==0.12.2
    # via -r requirements.in
semver==2.13.0
    # via streamlit
six==1.16.0
    # via
    #   astunparse
    #   google-auth
    #   google-pasta
    #   python-dateutil
    #   tensorflow
sklearn==0.0.post1
    # via -r requirements.in
smmap==5.0.0
    # via gitdb
streamlit==1.20.0
    # via -r requirements.in
tensorboard==2.11.2
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.11.0
    # via -r requirements.in
tensorflow-estimator==2.11.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.31.0
    # via tensorflow
termcolor==2.2.0
    # via tensorflow
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.2
    # via streamlit
tqdm==4.65.0
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   streamlit
    #   tensorflow
tzdata==2022.7
    # via pytz-deprecation-shim
tzlocal==4.2
    # via streamlit
urllib3==1.26.14
    # via requests
validators==0.20.0
    # via streamlit
watchdog==2.3.1
    # via streamlit
werkzeug==2.2.3
    # via tensorboard
wheel==0.38.4
    # via
    #   astunparse
    #   tensorboard
wrapt==1.15.0
    # via tensorflow
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
